Remove Pokemon Decrypt

+++ include/random.h
@@ -11,7 +11,7 @@ u16 Random(void);
 u16 Random2(void);
 
 //Returns a 32-bit pseudorandom number
-#define Random32() (Random() | (Random() << 16))
+#define Random32() ((u32)(Random() | (Random() << 16)))
 
 // The number 1103515245 comes from the example implementation of rand and srand
 // in the ISO C standard.
